% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dyslexic.R
\name{theme_dyslexic}
\alias{theme_dyslexic}
\title{ggplot2 theme with OpenDyslexic font}
\usage{
theme_dyslexic(base_size = 14)
}
\arguments{
\item{base_size}{Numeric. Base font size.}
}
\value{
A ggplot2 theme object.
}
\description{
ggplot2 theme with OpenDyslexic font
}
