% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dyslexic.R
\name{use_dyslexic}
\alias{use_dyslexic}
\title{Load and/or Activate OpenDyslexic for ggplot2}
\usage{
use_dyslexic(active = TRUE, load_font = TRUE)
}
\arguments{
\item{active}{Logical. If TRUE (default), sets OpenDyslexic as the global
theme and geom default. Requires load_font to be TRUE or already loaded.}

\item{load_font}{Logical. If TRUE (default), registers the OpenDyslexic
font files with R's font library.}
}
\description{
Load and/or Activate OpenDyslexic for ggplot2
}
