% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dyslexic.R
\name{use_dyslexic}
\alias{use_dyslexic}
\title{Register and enable OpenDyslexic for ggplot2}
\usage{
use_dyslexic()
}
\description{
This function locates the OpenDyslexic font files shipped with the package,
registers them with the system, and enables showtext rendering.
}
\examples{
\dontrun{
library(ggplot2)
use_dyslexic()
ggplot(mtcars, aes(hp, mpg)) + geom_point() + theme_dyslexic()
}
}
